+++ src/main/java/io/stargate/sgv2/jsonapi/service/cqldriver/executor/NamespaceCache.java
@@ -113,4 +113,8 @@ public class NamespaceCache {
               }
             });
   }
+
+  public void evictCollectionSettingCacheEntry(String collectionName) {
+    vectorCache.invalidate(collectionName);
+  }
 }